backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --system requirements.in -o requirements.txt
alembic==1.12.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via pwdlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.1.2
    # via
    #   passlib
    #   pwdlib
cachetools==5.5.0
    # via premailer
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
chardet==5.2.0
    # via emails
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via uvicorn
cryptography==43.0.3
    # via
    #   pyjwt
    #   python-jose
cssselect==1.2.0
    # via premailer
cssutils==2.11.1
    # via
    #   emails
    #   premailer
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.1.2
    # via
    #   fastapi-users
    #   pydantic
emails==0.6
    # via -r requirements.in
fastapi==0.115.2
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-users==13.0.0
    # via
    #   -r requirements.in
    #   fastapi-users-db-sqlalchemy
fastapi-users-db-sqlalchemy==6.0.1
    # via
    #   -r requirements.in
    #   fastapi-users
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   requests
jinja2==3.1.4
    # via -r requirements.in
lxml==5.3.0
    # via
    #   emails
    #   premailer
makefun==1.15.6
    # via fastapi-users
mako==1.3.5
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
more-itertools==10.5.0
    # via cssutils
passlib==1.7.4
    # via -r requirements.in
premailer==3.10.0
    # via emails
psycopg2-binary==2.9.10
    # via -r requirements.in
pwdlib==0.2.0
    # via fastapi-users
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via
    #   -r requirements.in
    #   pydantic
pydantic-settings==2.6.0
    # via -r requirements.in
pyhumps==3.8.0
    # via -r requirements.in
pyjwt==2.8.0
    # via fastapi-users
python-dateutil==2.9.0.post0
    # via emails
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via
    #   -r requirements.in
    #   fastapi-users
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   emails
    #   premailer
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
    #   fastapi-users-db-sqlalchemy
starlette==0.40.0
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.3
    # via requests
uvicorn==0.23.2
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
